+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-beta.280
+
+### Patch Changes
+
+- Updated dependencies [405fa0f]
+  - sit-onyx@1.0.0-beta.283
+
 ## 1.0.0-beta.279
 
 ### Patch Changes